requirements.txt
pyqtdarktheme==1.1.0
pyserial==3.5
PySide6==6.3.0
PySide6-Addons==6.3.0
PySide6-Essentials==6.3.0
shiboken6==6.3.0
